- Add sensor wrapper for light and environment sensor - Replace include guards of all sensor wrapper with "pragma once" - Fix missing light sensor output in "Sensors Summary" app - Set value for light sensor and IAQ to -1 when the sensors aren´t available in "Sensors Summary" app - Swap "Rel. height" and "IAQ" in "Sensors Summary" app

+++ app/src/applications/iaq/ui_export/screens/iaq_ui.c
@@ -128,7 +128,7 @@ void iaq_app_ui_remove(void)
     ui_root_page = NULL;
 }
 
-void iaq_app_ui_home_set_iaq_cursor(uint16_t iaq)
+void iaq_app_ui_home_set_iaq_cursor(float iaq)
 {
     int16_t angle;
 
@@ -146,12 +146,12 @@ void iaq_app_ui_home_set_iaq_cursor(uint16_t iaq)
     lv_img_set_angle(ui_ImgCursor, angle);
 }
 
-void iaq_app_ui_home_set_iaq_label(int16_t iaq)
+void iaq_app_ui_home_set_iaq_label(float iaq)
 {
     lv_label_set_text_fmt(ui_LabelAQI, "%u", iaq);
 }
 
-void iaq_app_ui_home_set_iaq_status(int16_t iaq)
+void iaq_app_ui_home_set_iaq_status(float iaq)
 {
     char* text;
     lv_color_t color;
